REF: wait() moved out of statusbase
